changes 'bash_completion' to 'bash_scripts'

+++ bash_profile
@@ -1,23 +1,29 @@
 #!/bin/bash
 
-if [ -f $HOME/dotfiles/bashrc ]; then
+if [ -f $HOME/dotfiles/bashrc ] && [ ! -f $HOME/.bashrc ]; then
   source "$HOME/dotfiles/bashrc"
 fi
 
-if [ -f $HOME/dotfiles/bash_aliases ]; then
+if [ -f $HOME/dotfiles/bash_aliases ] && [ ! -f $HOME/.bash_aliases ]; then
   source "$HOME/dotfiles/bash_aliases"
 fi
 
-if [ ! -f $HOME/.tmux.conf ]; then
+# if [ -d $HOME/dotfiles/scripts ] && [ ! -d $HOME/.scripts ]; then
+#   ln -s "$HOME/dotfiles/scripts $HOME/.scripts"
+# fi
+
+if [ -f $HOME/dotfiles/tmux.conf ] && [ ! -f $HOME/.tmux.conf ]; then
   ln -s "$HOME/dotfiles/tmux.conf" "$HOME/.tmux.conf"
 fi
 
-if [ ! -f $HOME/.Xresources ]; then
-  ln -s "$HOME/dotfiles/Xresources" "$HOME/.Xresources"
+if [ -f $HOME/dotfiles/inputrc ] && [ ! -f $HOME/.inputrc ]; then
+  ln -s "$HOME/dotfiles/inputrc" "$HOME/.inputrc"
 fi
 
-if [ ! -f $HOME/.inputrc ]; then
-  ln -s "$HOME/dotfiles/inputrc" "$HOME/.inputrc"
+if [[ "$OSTYPE" == "linux-gnu" ]]; then
+  if [ -x "$(command -v Xorg)" ] && [ ! -f $HOME/.Xresources ]; then
+    ln -s "$HOME/dotfiles/Xresources" "$HOME/.Xresources"
+  fi
 fi
 
 # if [ ! -f $HOME/.vimrc ]; then
